telegram-bot/requirements.txt
python-telegram-bot==13.1
bs4
requests
psycopg2
uvicorn
fastapi
pydantic
